Migrate Kotlin DSL IDE Settings script template to new API
